+++ hosts/phi/hardware-configuration.nix
@@ -31,7 +31,7 @@
     };
 
   fileSystems."/data" =
-    { device = "/dev/disk/by-uuid/a0346b3d-e7b2-4b77-bd3e-f06859580c82";
+    { device = "/dev/disk/by-label/data";
       fsType = "ext4";
     };